These are the updated files as of 9/10/25
QBar works when on and Q off, but the other way round does not work. This can be seen in the files above. Also the values for the LC filter are done, the >= 1A specification has not been met yet. This is the next step after figuring out why Q on and QBar off is not working as thought to be.